DATAMONGO-941 - Add support for Update $min.
We now offer dedicated methods for ‘min’ (Number/Date) on Update.
As BigInteger and BigDecimal values are converted to String before storing in MongoDB it is not possible to use for the min operation. Therefore we type check both the given value and the property it should be applied to and raise an error if  it fails.